Extract craft command class to a separate module.
