+++ CataclysmWin.cbp
@@ -370,6 +370,8 @@
 		<Unit filename="src/construction.cpp" />
 		<Unit filename="src/construction.h" />
 		<Unit filename="src/coordinates.h" />
+		<Unit filename="src/craft_command.cpp" />
+		<Unit filename="src/craft_command.h" />
 		<Unit filename="src/crafting.cpp" />
 		<Unit filename="src/crafting.h" />
 		<Unit filename="src/creature.cpp" />
