+++ src/player.h
@@ -2,11 +2,12 @@
 #define PLAYER_H
 
 #include "character.h"
+#include "crafting.h"
+#include "craft_command.h"
 #include "item.h"
 #include "player_activity.h"
-#include "weighted_list.h"
 #include "morale.h"
-#include "crafting.h"
+#include "weighted_list.h"
 
 #include <unordered_set>
 #include <bitset>
@@ -938,15 +939,15 @@ class player : public Character, public JsonSerializer, public JsonDeserializer
         const inventory &crafting_inventory(); // includes nearby items
         void invalidate_crafting_inventory();
         std::vector<item> get_eligible_containers_for_crafting();
-        comp_selection<item_comp> select_item_component( const std::vector<item_comp> &components,
-                                                         int batch, inventory &map_inv,
-                                                         bool can_cancel = false );
+        comp_selection<item_comp>
+            select_item_component( const std::vector<item_comp> &components,
+                                   int batch, inventory &map_inv, bool can_cancel = false );
         std::list<item> consume_items( const comp_selection<item_comp> &cs, int batch );
         std::list<item> consume_items( const std::vector<item_comp> &components, int batch = 1 );
-        comp_selection<tool_comp> select_tool_component( const std::vector<tool_comp> &tools,
-                                                         int batch, inventory &map_inv,
-                                                         const std::string &hotkeys = DEFAULT_HOTKEYS,
-                                                         bool can_cancel = false );
+        comp_selection<tool_comp>
+            select_tool_component( const std::vector<tool_comp> &tools, int batch, inventory &map_inv,
+                                   const std::string &hotkeys = DEFAULT_HOTKEYS,
+                                   bool can_cancel = false );
         void consume_tools( const comp_selection<tool_comp> &tool, int batch );
         void consume_tools( const std::vector<tool_comp> &tools, int batch = 1,
                             const std::string &hotkeys = DEFAULT_HOTKEYS );
